webdav: Do not throw when saving new document.
Propagate information about !shouldAccessNetworkAfterException
from getResourceType out to getPropertyValues, to avoid further calls
which would result in cancelCommandExecution, when trying get properties
of document which does not exist yet, but we want to create it.

This was done in webdav-neon for other reasons as part of
d89c6228279d9f5adfee04259b37cc8e3c8468e3.

Change-Id: Ia434a707d2946a6982075f9bf53904b78dc07a5e