fdo#78758: sw: RTF export: don't export multiple \fldrst for one hyperlink
Ensure that we export only one \fldresult per hyperlink by doing that in
StartURL() and EndURL(); the TextINetFormat() is called once per text
portion.  This shouldn't cause problems as there can't be anything
between the end of the \field group and the start of \fldresult anyway.

Replace the annoying call to EndURL() from EndRun() with a special case
in EndURL() to store things in the right buffer (hopefully).

(somehow this is regression from 39031a5d54a25b6730ca15addb2664d72ba3f062)

Change-Id: I209ea7a384fb1cb5d1505a70ecc4a4536bbf26a2