Related: bnc#875718 DOCX import: fix missing character grab-bag on fields
The problem was that in case:

1) The paragraph only had a single text portion, which was a field and
char grab-bag was set on it and
2) The paragraph had a style which also set the character grab-bag

then during import the field's gra-bag was set on the paragraph (as it's
the only portion) and later the paragraph style overwrote this, in case
that had a grab-bag, too.

Work this around by ensuring that in case of portion fields (i.e. not
ToC, which has its own paragraphs), there are always at least two
portions in a paragraph (the second is removed later).

This also fixes the fake paragraph problem at the end of the bnc#875718
testcase. (There was an empty paragraph at the end of the document, but
not in the file itself.)

Change-Id: Ie404bc043d46157ea6157b18c4a46395cf496118