fdo#80924: writerfilter: RTF import: fix shapeType PictureFrame import
For shapeType PictureFrame, a default CustomShape was created and then
RTFDocumentImpl::resolvePict() crashes because it actually operates on a
previous shape, because in this special case RTFSdrImport::m_xShape
is never actually set to the new shape, so contains the previous one.

Also the new shape needs to be added to the draw-page, at least
otherwise the assertion of the supported service fails because
some SvxShape::mpObj weak-reference is dead?

This essentially reverts commit 6805bae83d32c36cfdb0239ec08a15608c7c5196
(except for the added testcase, which was slightly different because
it did not have a "pib" property).

(regression from f796704b0871a9e3e8a4d0149ea7094550560aea)

Change-Id: I6539c4286850dff2d8564006487cc765f1d117be