DOCX filter: map <w:sdt> outside paragraphs to paragraph properties
Previously the type of the SDT was used to decide if the SDT will be a
character or paragraph property. Improve this situation by always
mapping <w:sdt> elements outside paragraph to paragraph properties.

In practice, this means that if the SDT was a rectangle (as it wasn't
only around a run), then it remains so, while previously it could become
a polygon instead (when painted).

Fix several testcases that tested that a <w:sdt> outside a paragraph is
exported as a character property.

Change-Id: Ia26c1a4cf6bc553b46224e4b17ee40725c5f3622