suppress legacy locales, fdo#44112 related
With MsLangId::isLegacy() locales can be suppressed in language lists
and are not offered unless actually used at the current position of the
document. Language attribution now handles adding a language/locale to
the list on the fly.

For example, Serbian (Latin or Cyrillic) for "Serbia and Montenegro" are
legacy locales, the confederation doesn't exist anymore. A further
candidate would be pap-AN as the Netherlands Antilles ceased to exist,
see fdo#44112.

Suppressing "Serbian Cyrillic (Serbia and Montenegro)" also has the nice
benefit of narrowing the language list box width a little ...

Change-Id: Id05cce328622465016901524b759369c18bc4dbd