fdo#78611 Don't create automatic rules on reset
As we reset all styles, there shouldn't be any unknown automatic
SwNumRules, because all should have been created by the style
copying! And in this case we also have to overwrite the existing
automatic SwNumRules to get a consistent copy.

In case of unknown automatic SwNumRules we just warn and ignore
any inconsistence.

Change-Id: I26e25bccc9a8c7be74f3dceb2e1f91894eac1d0a
Reviewed-on: https://gerrit.libreoffice.org/10988
Reviewed-by: Björn Michaelsen <bjoern.michaelsen@canonical.com>
Tested-by: Björn Michaelsen <bjoern.michaelsen@canonical.com>