"fixes" needed because ure/source/unorc is now expanded at configure time
Argh, this is getting even uglier.

We cheerfully ignore for now the theoretical possibility that the URE
unorc used by build-time tools (i.e. the configure-expanded
ure/source/unorc) could be different for HOST and BUILD (in case they
use different --enable-canonical-installation-tree-structure), and use
the HOST one for the BUILD tools.

The right thing would probably be to construct the URE unorc in the
relevant Makefile, like we do for fundamentalrc? Or then to just
re-design the whole mess of rc files into some simpler (good luck).

Change-Id: I654309503d0e696778910acadcbf2f6b90ffa02a