fdo#48110 disable the "Automatically" currency unless already used
This disables the "Automatically" currency listbox entry in the number
formats dialog and makes it unselectable, unless such currency format
was already used at the current position of a loaded document.

Getting completely rid of the listbox entry if not used is more
cumbersome as the positions are tied to the list entries in
SvxNumberFormatShell::aCurCurrencyList, but at least the disabled entry
prevents accidental or ill intended selection of that awkward format.

Change-Id: I16f3e88feede34e3f83afb108f253fb53d82d34d