writerfilter: abort build if model.xml does not validate
Most of the scripts parsing model.xml just load it as a DOM and extract
info from it needed for their task, which means if you have a typo in an
element / attribute name, it's just silently ignored.

This is annoying, let's prevent that by validating the model before
doing anything with it.

Change-Id: I4229e57c12840fead62cbd150dd1408f2f0d29eb