split CppunitTest_sw_ooxmlexport into several tests
In order to allow running them in parallel. There are already enough
tests to make it run for a couple of minutes and it's annoying to
see only once core busy with it.
I intentionally left the original file at the same name, without 1
appended, in order to avoid merge problems.

Change-Id: I097c45c34797cc68f5d1790cf31eefc96fbaf44e